requirements.txt
# GN - SLA April 2021
# From 'pip freeze' on STG site
argparse==1.3.0
beautifulsoup4==4.4.0
bleach==1.2.2
chardet==2.3.0
Django==1.5.4
django-contrib-comments==1.6.1
django-debug-toolbar==1.2.2
django-grappelli==2.4.5
django-haystack==2.4.0
django-modeltranslation==0.8
django-redis-cache==0.13.0
# No longer used?
# django-smart-selects==1.0.8
elasticsearch==1.6.0
filebrowser-safe==0.4.0
future==0.14.3
# https://stackoverflow.com/a/17311033
# sudo add-apt-repository ppa:ubuntugis/ppa
# GDAL==1.6.3
geopy==0.95.1
grappelli-safe==0.4.0
hiredis==0.1.5
html5lib==0.95
importlib==1.0.3
linecache2==1.0.0
Mezzanine==1.4.7
mpmath==0.19
numpy==1.4.1
oauthlib==0.7.2
ordereddict==1.1
pbr==1.1.1
Pillow==2.0.0
# psycopg2==2.5
psycopg2-binary<3
pyparsing==1.5.2
python-dateutil==1.4.1
pytz==2013b0
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
Shapely==1.2.1
six==1.9.0
sorl-thumbnail==12.3
South==1.0.2
sqlparse==0.1.14
traceback2==1.4.0
tzlocal==1.2
unittest2==1.0.1
urllib3==1.10.4
virtualenv==13.1.0
wheel==0.24.0
# Unused (ES)
# Whoosh==2.7.0
# no longer used (ES)
# xapian-haystack==2.0.0
